feat: improve repository card hover footer
* move repository page helpers into src/helpers
* make card footers shift to stronger colors on hover and focus
