+++ src/components/repositoryPageContent/content.tsx
@@ -1,10 +1,11 @@
 'use client';
 
-import { CSSProperties, useState } from 'react';
+import { useState } from 'react';
 
-import { Colorscheme } from '@/models/colorscheme';
 import { RepositoryDTO } from '@/models/DTO/repository';
 import { Repository } from '@/models/repository';
+
+import { RepositoryPageHelper } from '@/helpers/repositoryPage';
 
 import RepositoryInfo from '@/components/repositoryInfo/repositoryInfo';
 import RepositoryTitle from '@/components/repositoryTitle';
@@ -18,19 +19,6 @@ type RepositoryPageContentClientProps = {
   repositoryDTO: RepositoryDTO;
 };
 
-const swatchGroupPriority = [
-  'NormalBg',
-  'NormalFg',
-  'vimLineCommentFg',
-  'vimStringFg',
-  'vimFuncNameFg',
-  'vimNumberFg',
-  'vimCommandFg',
-  'StatusLineBg',
-  'CursorLineBg',
-  'LineNrFg',
-];
-
 export default function RepositoryPageContentClient({
   repositoryDTO,
 }: RepositoryPageContentClientProps) {
@@ -38,8 +26,9 @@ export default function RepositoryPageContentClient({
   const variants = repository.flattenedColorschemes;
   const [activeIndex, setActiveIndex] = useState(0);
   const activeVariant = variants[activeIndex];
-  const colorschemeStyle = getColorschemeStyle(activeVariant);
-  const swatchColors = getSwatchColors(activeVariant);
+  const colorschemeStyle =
+    RepositoryPageHelper.getColorschemeStyle(activeVariant);
+  const swatchColors = RepositoryPageHelper.getSwatchColors(activeVariant);
 
   return (
     <div className={styles.layout} style={colorschemeStyle}>
@@ -102,41 +91,5 @@ export default function RepositoryPageContentClient({
         )}
       </TuiSection>
     </div>
-  );
-}
-
-function getSwatchColors(colorscheme: Colorscheme | undefined): string[] {
-  const background = colorscheme?.backgrounds[0];
-
-  if (!colorscheme || !background) {
-    return [];
-  }
-
-  const groups = colorscheme.data[background] ?? [];
-  const colorByGroup = new Map(groups.map(group => [group.name, group.hexCode]));
-  const prioritizedColors = swatchGroupPriority
-    .map(groupName => colorByGroup.get(groupName))
-    .filter((color): color is string => Boolean(color));
-
-  return Array.from(
-    new Set([...prioritizedColors, ...groups.map(group => group.hexCode)]),
-  ).slice(0, 6);
-}
-
-function getColorschemeStyle(
-  colorscheme: Colorscheme | undefined,
-): CSSProperties | undefined {
-  const background = colorscheme?.backgrounds[0];
-
-  if (!colorscheme || !background) {
-    return undefined;
-  }
-
-  return colorscheme.data[background]?.reduce<CSSProperties>(
-    (acc, group) => ({
-      ...acc,
-      [`--colorscheme-${group.name}`]: group.hexCode,
-    }),
-    {},
   );
 }
